Add jupyter as a dev tool
By following the instructions on the README.md, I could correctly start
the virtual environment constructed by poetry and launch the jupyter
notebooks.

However, for some reason all notebooks detected that I had installed the
dependencies, but couldn't import any of them. I suspected that this was
something related to the jupyter notebook not recognizing the virtual
environment somehow, so I fixed the issue by adding jupyter as a poetry
dependency and running this virtual version (instead of the local one
installed on the host).

In this sense, add jupyter as a developer dependency to avoid the
unrecognized virtual environment.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>